+++ src/components/Admin/ManageSensors/DownloadData/Util.tsx
@@ -39,13 +39,13 @@ interface CSVButtonProps {
   endDay: number;
   error: string;
   downloadAll: boolean;
-  paId: string;
+  purpleAirId: string;
   resetSelectedSensor: () => void;
 }
 
 /**
  * Props for the CSV Modal.
- * sensors - the list of PA sensors to include in the drop down menu
+ * sensors - the list of PurpleAir sensors to include in the drop down menu
  */
 interface CSVModalProps {
   sensors: Sensor[];
@@ -58,64 +58,6 @@ interface InputProps {
   value: number;
   setValue: React.Dispatch<React.SetStateAction<number>>;
 }
-
-/**
- * Props for sensor input fields used in the `DownloadCSVModal`
- */
-interface SensorInputProps {
-  sensors: Sensor[];
-  value: string;
-  setValue: React.Dispatch<React.SetStateAction<string>>;
-}
-
-/**
- * Drop-down menu for the sensors in our database. This is used in the `DownloadCSVModal`
- * when downloading data for a single sensor instead of multiple sensors
- * @param sensors - the list of sensors to choose from
- * @param value - the value (state variable) that should be updated by this drop-down
- * @param setValue - a function that sets the state of the value
- * @returns a drop down menu with all the sensors
- */
-const SensorInput: ({
-  sensors,
-  value,
-  setValue,
-}: SensorInputProps) => JSX.Element = ({
-  sensors,
-  value,
-  setValue,
-}: SensorInputProps) => {
-  const {t} = useTranslation('administration');
-  const options = [];
-  for (let i = 0; i < sensors.length; i++) {
-    const sensor = sensors[i];
-    let label;
-    if (sensor.name) {
-      label = sensor.name;
-    } else {
-      label = sensor.purpleAirId;
-    }
-    options.push(
-      <option value={sensor.purpleAirId} key={i}>
-        {label}
-      </option>
-    );
-  }
-  return (
-    <Box>
-      <Select
-        type="number"
-        placeholder={t('downloadData.chooseSensor')}
-        value={value}
-        onChange={event => {
-          setValue(event.target.value);
-        }}
-      >
-        {options}
-      </Select>
-    </Box>
-  );
-};
 
 /**
  * Drop-down menu for months of the year
@@ -208,4 +150,4 @@ export type {
   InputProps,
   CSVModalProps,
 };
-export {MonthInput, DayInput, SensorInput};
+export {MonthInput, DayInput};
